# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Using a 1C cluster control service over http",
        "title": "1C cluster control service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/cluster/:cluster/connection/list": {
            "get": {
                "description": "Show all connections with identifiers for current cluster",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connection list"
                ],
                "summary": "Show all connections in cluster",
                "operationId": "connections",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of cluster",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Firstly try to find from Cache",
                        "name": "cache",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Entrypoint for cluster",
                        "name": "entrypoint",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.connectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    }
                }
            }
        },
        "/cluster/:cluster/infobase/:infobase/connection/list": {
            "get": {
                "description": "Show all connections with identifiers for current infobase in cluster",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "connection list infobase"
                ],
                "summary": "Show all connections in infobase",
                "operationId": "connectionsByInfobase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of cluster",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID of infobase",
                        "name": "infobase",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Firstly try to find from Cache",
                        "name": "cache",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Entrypoint for cluster",
                        "name": "entrypoint",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.connectionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    }
                }
            }
        },
        "/cluster/:cluster/infobase/:infobase/session/list": {
            "get": {
                "description": "Show all sessions with identifiers for current infobase in cluster",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session list infobase"
                ],
                "summary": "Show all sessions in infobase",
                "operationId": "sessionsByInfobase",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of cluster",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UUID of infobase",
                        "name": "infobase",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Firstly try to find from Cache",
                        "name": "cache",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Entrypoint for cluster",
                        "name": "entrypoint",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.sessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    }
                }
            }
        },
        "/cluster/:cluster/infobase/list": {
            "get": {
                "description": "Show all infobases with identifiers for current cluster",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "infobase list"
                ],
                "summary": "Show all infobases in cluster",
                "operationId": "infobases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of cluster",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Firstly try to find from Cache",
                        "name": "cache",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Entrypoint for cluster",
                        "name": "entrypoint",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.infobaseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    }
                }
            }
        },
        "/cluster/:cluster/session/list": {
            "get": {
                "description": "Show all sessions with identifiers for current cluster",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "session list"
                ],
                "summary": "Show all sessions in cluster",
                "operationId": "sessions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of cluster",
                        "name": "cluster",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Firstly try to find from Cache",
                        "name": "cache",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Entrypoint for cluster",
                        "name": "entrypoint",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.sessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    }
                }
            }
        },
        "/cluster/list": {
            "get": {
                "description": "Show all clusters with data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster list"
                ],
                "summary": "Show clusters",
                "operationId": "clusters",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Firstly try to find from Cache",
                        "name": "cache",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Entrypoint for cluster",
                        "name": "entrypoint",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.clusterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/error.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Cluster": {
            "type": "object",
            "properties": {
                "errth": {
                    "type": "integer",
                    "example": 0
                },
                "exp": {
                    "type": "integer",
                    "example": 1200
                },
                "host": {
                    "type": "string",
                    "example": "localhost"
                },
                "id": {
                    "type": "string",
                    "example": "UUID"
                },
                "kpp": {
                    "type": "integer",
                    "example": 0
                },
                "lb": {
                    "type": "string",
                    "example": "perfomance"
                },
                "lt": {
                    "type": "integer",
                    "example": 3000
                },
                "mms": {
                    "type": "integer",
                    "example": 50000
                },
                "mmts": {
                    "type": "integer",
                    "example": 600
                },
                "name": {
                    "type": "string",
                    "example": "name"
                },
                "port": {
                    "type": "string",
                    "example": "1541"
                },
                "sftl": {
                    "type": "integer",
                    "example": 0
                },
                "sl": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "entity.Connection": {
            "type": "object",
            "properties": {
                "appid": {
                    "type": "string",
                    "example": "1CV8"
                },
                "blocked": {
                    "type": "integer",
                    "example": 0
                },
                "cid": {
                    "type": "integer",
                    "example": 12345
                },
                "connected": {
                    "type": "string",
                    "example": "2023-08-10T11:40:55"
                },
                "host": {
                    "type": "string",
                    "example": "localhost"
                },
                "ib": {
                    "type": "string",
                    "example": "UUID"
                },
                "id": {
                    "type": "string",
                    "example": "UUID"
                },
                "proc": {
                    "type": "string",
                    "example": "UUID"
                },
                "sid": {
                    "type": "integer",
                    "example": 12345
                }
            }
        },
        "entity.Infobase": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string",
                    "example": "comments"
                },
                "id": {
                    "type": "string",
                    "example": "UUID"
                },
                "name": {
                    "type": "string",
                    "example": "name"
                }
            }
        },
        "entity.Session": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "string",
                    "example": "2023-08-10T14:04:43"
                },
                "appid": {
                    "type": "string",
                    "example": "1CV8"
                },
                "blockdb": {
                    "type": "integer",
                    "example": 0
                },
                "blockls": {
                    "type": "integer",
                    "example": 0
                },
                "bytes": {
                    "type": "integer",
                    "example": 12345
                },
                "bytes5m": {
                    "type": "integer",
                    "example": 123
                },
                "bytesdb": {
                    "type": "integer",
                    "example": 123
                },
                "bytesdb5m": {
                    "type": "integer",
                    "example": 12
                },
                "calls": {
                    "type": "integer",
                    "example": 5
                },
                "calls5m": {
                    "type": "integer",
                    "example": 2
                },
                "conn": {
                    "type": "string",
                    "example": "UUID"
                },
                "cpu": {
                    "type": "integer",
                    "example": 1234
                },
                "cpu5m": {
                    "type": "integer",
                    "example": 12
                },
                "cpucur": {
                    "type": "integer",
                    "example": 123
                },
                "dbproc": {
                    "type": "integer",
                    "example": 123
                },
                "dbprocat": {
                    "type": "string",
                    "example": ""
                },
                "dbproci": {
                    "type": "string",
                    "example": ""
                },
                "dur": {
                    "type": "integer",
                    "example": 100
                },
                "dur5m": {
                    "type": "integer",
                    "example": 100
                },
                "durcur": {
                    "type": "integer",
                    "example": 80
                },
                "durcurdb": {
                    "type": "integer",
                    "example": 80
                },
                "durdb": {
                    "type": "integer",
                    "example": 100
                },
                "durdb5m": {
                    "type": "integer",
                    "example": 100
                },
                "dursvc": {
                    "type": "integer",
                    "example": 0
                },
                "dursvc5m": {
                    "type": "integer",
                    "example": 0
                },
                "dursvccur": {
                    "type": "integer",
                    "example": 0
                },
                "hib": {
                    "type": "string",
                    "example": "yes/no"
                },
                "hibterm": {
                    "type": "integer",
                    "example": 3600
                },
                "hibtm": {
                    "type": "integer",
                    "example": 1200
                },
                "host": {
                    "type": "string",
                    "example": "Host"
                },
                "ib": {
                    "type": "string",
                    "example": "UUID"
                },
                "id": {
                    "type": "string",
                    "example": "UUID"
                },
                "loc": {
                    "type": "string",
                    "example": "ru"
                },
                "mem": {
                    "type": "integer",
                    "example": 123456
                },
                "mem5m": {
                    "type": "integer",
                    "example": 1234
                },
                "memcur": {
                    "type": "integer",
                    "example": 12345
                },
                "proc": {
                    "type": "string",
                    "example": "UUID"
                },
                "read": {
                    "type": "integer",
                    "example": 56789
                },
                "read5m": {
                    "type": "integer",
                    "example": 56
                },
                "readcur": {
                    "type": "integer",
                    "example": 5678
                },
                "sep": {
                    "type": "string",
                    "example": ""
                },
                "sid": {
                    "type": "integer",
                    "example": 12345
                },
                "started": {
                    "type": "string",
                    "example": "2023-08-10T14:04:43"
                },
                "svc": {
                    "type": "string",
                    "example": "Name"
                },
                "uname": {
                    "type": "string",
                    "example": "UserName"
                },
                "write": {
                    "type": "integer",
                    "example": 123
                },
                "write5m": {
                    "type": "integer",
                    "example": 123
                },
                "writecur": {
                    "type": "integer",
                    "example": 123
                }
            }
        },
        "error.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "v1.clusterResponse": {
            "type": "object",
            "properties": {
                "clusters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Cluster"
                    }
                }
            }
        },
        "v1.connectionResponse": {
            "type": "object",
            "properties": {
                "connections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Connection"
                    }
                }
            }
        },
        "v1.infobaseResponse": {
            "type": "object",
            "properties": {
                "infobases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Infobase"
                    }
                }
            }
        },
        "v1.sessionResponse": {
            "type": "object",
            "properties": {
                "sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Session"
                    }
                }
            }
        }
    }
}